src/requirements.txt
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
transformers==4.35.0
torch==2.1.0
yfinance==0.2.28
newsapi-python==0.2.7
ta==0.11.0
shap==0.43.0
mlflow==2.8.0
optuna==3.4.0
fastapi==0.104.0
uvicorn==0.24.0
streamlit==1.28.0
plotly==5.18.0
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0